Update GuessingGame.md
The game works by having the computer randomly select a number between 0 and 20. The formula is n-1 so whatever number we pick, the guessing range will be one lower, in this case we selected 21 to make the range 0-20. Then as the user guesses a number, the program is designed to provide you with the hint Higher or Lower based on your guess. Once the correct guess has been made, the program will say "Congrats! That's Correct. Thank you for playing" letting you know you've won